reorganized test locations in folders
