+++ scout/tests.py
@@ -7,6 +7,7 @@ ACCESS_KEY = getattr(settings, 'SAUCE_ACCESS_KEY', False)
 
 if USERNAME and ACCESS_KEY:
     from scout.test.navigation import NavigationTest
-    from scout.test.urlStatus import UrlStatusTest
+    from scout.test.pageflow.navigation import MainNavigationTest
+    from scout.test.pageflow.pageloadStatus import UrlStatusTest
     from scout.test.fidelity_ui import UITest
-    from scout.test.wireframe_navigation import WireframeTest
+    from scout.test.wireframe.wireframe_navigation import WireframeTest
